backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
scikit-learn>=1.5.0
pandas>=2.1.2
numpy>=1.26.0
joblib>=1.3.2
requests>=2.31.0
python-whois>=0.8.0
python-dotenv>=1.0.0
